rename and lint frontend
